USER_C_MODULES cmake support for rp2 and esp32
This change introduces `usermod.cmake`, the CMake configuration required to make
this library compatible with the CMake build systems for the rp2 (Raspberry Pi Pico/RP2040)
and esp32 ports of MicroPython.

This is effectively 1:1 in functionality with `micropython.mk` and defines the source files,
include directories and compile arguments needed to build the module.

See https://github.com/micropython/micropython/pull/6960 for discussion.